Protect edit

+++ controllers/videoController.js
@@ -40,9 +40,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
-  console.log(newVideo);
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 }
   
@@ -51,7 +53,7 @@ export const videoDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate('creator');
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -63,8 +65,12 @@ export const getEditVideo = async (req, res) => {
   // 찾아낸 아이디로 database에서 정보를 title과 description의 정보를 가져온다.
   const { params: { id } } = req;
   try {
-    const video = await Video.findById(id)
-    res.render("editVideo", { pageTitle: "Edit Video", video });
+    const video = await Video.findById(id);
+    if (video.creator.id !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -92,7 +98,11 @@ export const deleteVideo = async (req, res) => {
     }
   } = req;
   try {
-    await Video.findOneAndRemove({_id: id});
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({_id: id});
+    }
     res.redirect(routes.home);
   } catch (error) {
     res.redirect(routes.home);
